Use Pkg extensions
