+++ Project.toml
@@ -64,6 +64,23 @@ ValueShapes = "136a8f8c-c49b-4edb-8b98-f3d64d48be8f"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 ZygoteRules = "700de1a5-db45-46bc-99cf-38207098b444"
 
+[weakdeps]
+Cuba = "8a292aeb-7a57-582c-b821-06e4c11590b1"
+Folds = "41a02a25-b8f0-4f67-bc48-60067656b558" 
+HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
+NestedSamplers = "41ceaf6f-1696-4a54-9b49-2e7a9ec3782e"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+Transducers = "28d57a85-8fef-5791-bfe6-a80928e7c999"
+UltraNest = "6822f173-b0be-4018-9ee2-28bf56348d09"
+
+[extensions]
+BATCubaExt = "Cuba"
+BATFoldsExt = ["Folds", "Transducers"]
+BATHDF5Ext = "HDF5"
+BATNestedSamplersExt = "NestedSamplers"
+BATPlotsExt = "Plots"
+BATUltraNestExt = "UltraNest"
+
 [compat]
 AdvancedHMC = "0.3, 0.4"
 ArgCheck = "1, 2.0"
@@ -72,6 +89,7 @@ ChainRulesCore = "0.9.44, 0.10, 1"
 ChangesOfVariables = "0.1.1"
 Clustering = "0.13, 0.14, 0.15"
 Colors = "0.9,0.10,0.11, 0.12"
+Cuba = "2"
 DensityInterface = "0.4"
 DiffResults = "1"
 Distributions = "0.25"
@@ -83,10 +101,12 @@ EmpiricalDistributions = "0.2, 0.3.1"
 FFTW = "1"
 FillArrays = "0.7 - 1.0.0, 1.1.1"
 FiniteDiff = "2.8"
+Folds = "0.2" 
 ForwardDiff = "0.10"
 ForwardDiffPullbacks = "0.1.1, 0.2"
 FunctionChains = "0.1"
 Functors = "0.2, 0.3, 0.4"
+HDF5 = "0.15, 0.16"
 HypothesisTests = "0.10, 0.11"
 IntervalSets = "0.5, 0.6, 0.7"
 InverseFunctions = "0.1"
@@ -96,12 +116,14 @@ LaTeXStrings = "1"
 MacroTools = "0.5"
 MeasureBase = "0.12, 0.13, 0.14"
 Measurements = "2"
-NLSolversBase = "7"
 NamedArrays = "0.9"
+NestedSamplers = "0.8"
+NLSolversBase = "7"
 Optim = "0.19,0.20, 0.21, 0.22, 1"
-PDMats = "0.9, 0.10, 0.11"
 ParallelProcessingTools = "0.4"
 Parameters = "0.12"
+PDMats = "0.9, 0.10, 0.11"
+Plots = "1"
 PositiveFactorizations = "0.2"
 PrettyTables = "1, 2"
 Random123 = "1.2"
@@ -114,8 +136,19 @@ StatsBase = "0.32, 0.33, 0.34"
 StatsFuns = "1"
 StructArrays = "0.4, 0.5, 0.6"
 Tables = "0.2, 1.0"
+Transducers = "0.4"
 TypedTables = "1.2"
+UltraNest = "0.1"
 ValueShapes = "0.10.1"
 Zygote = "0.6"
 ZygoteRules = "0.2"
 julia = "1.6"
+
+[extras]
+Cuba = "8a292aeb-7a57-582c-b821-06e4c11590b1"
+Folds = "41a02a25-b8f0-4f67-bc48-60067656b558" 
+HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
+NestedSamplers = "41ceaf6f-1696-4a54-9b49-2e7a9ec3782e"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+Transducers = "28d57a85-8fef-5791-bfe6-a80928e7c999"
+UltraNest = "6822f173-b0be-4018-9ee2-28bf56348d09"
