+++ docs/make_notutorial.jl
@@ -9,7 +9,6 @@ using Literate
 using BAT
 
 using DensityInterface, ValueShapes
-import Cuba, #=UltraNest,=# NestedSamplers
 
 # Import HDF5 only on 64-bit, automated installation of HDF5 doesn't seem to work
 # properly on 32-bit on CI system:
