+++ docs/src/stable_api.md
@@ -65,20 +65,23 @@ ARPWeighting
 AssumeConvergence
 AutocorLenAlgorithm
 bat_transform
+BATHDF5IO
+BATIOAlgorithm
 BrooksGelmanConvergence
 CholeskyPartialWhitening
 CholeskyWhitening
 DensitySample
 DensitySampleVector
 DistLikeMeasure
+DoNotTransform
 EffSampleSizeAlgorithm
 EffSampleSizeFromAC
 ExplicitInit
-IdentityTransformAlgorithm
 FullMeasureTransform
 GelmanRubinConvergence
 GeyerAutocorLen
 HamiltonianMC
+IdentityTransformAlgorithm
 IIDSampling
 InitFromIID
 InitFromSamples
@@ -86,9 +89,8 @@ InitFromTarget
 InitvalAlgorithm
 IntegrationAlgorithm
 KishESS
-LogDVal
 LBFGSOpt
-NelderMeadOpt
+LogDVal
 MaxDensitySearch
 MCMCAlgorithm
 MCMCBurninAlgorithm
@@ -102,7 +104,7 @@ MCMCTuningAlgorithm
 MetropolisHastings
 MHProposalDistTuning
 ModeAsDefined
-DoNotTransform
+NelderMeadOpt
 NoWhitening
 OrderedResampling
 PosteriorMeasure
