+++ src/BAT.jl
@@ -93,23 +93,26 @@ include("optimization/optimization.jl")
 include("samplers/samplers.jl")
 include("integration/integration.jl")
 include("algodefaults/algodefaults.jl")
-include("io/io.jl")
 include("plotting/plotting.jl")
+include("extdefs/extdefs.jl")
 include("deprecations.jl")
 
 # include("precompile.jl")
 
 
-const _PLOTS_MODULE = Ref{Union{Module,Nothing}}(nothing)
-_plots_module() = _PLOTS_MODULE[]
+@static if !isdefined(Base, :get_extension)
+    using Requires
+end
 
 function __init__()
-    @require Folds = "41a02a25-b8f0-4f67-bc48-60067656b558" @require Transducers = "28d57a85-8fef-5791-bfe6-a80928e7c999" include("utils/executors_folds.jl")
-    @require Cuba = "8a292aeb-7a57-582c-b821-06e4c11590b1" include("integration/cuba_integration.jl")
-    @require HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f" include("io/hdf5_specific.jl")
-    @require NestedSamplers = "41ceaf6f-1696-4a54-9b49-2e7a9ec3782e" include("samplers/nested_sampling/ellipsoidal_nested_sampling/ellipsoidal_nested_sampling.jl")
-    @require Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80" _PLOTS_MODULE[] = Plots
-    @require UltraNest = "6822f173-b0be-4018-9ee2-28bf56348d09" include("samplers/nested_sampling/ultranest.jl")
+    @static if !isdefined(Base, :get_extension)
+        @require Folds = "41a02a25-b8f0-4f67-bc48-60067656b558" @require Transducers = "28d57a85-8fef-5791-bfe6-a80928e7c999" include("../ext/BATFoldsExt.jl")
+        @require Cuba = "8a292aeb-7a57-582c-b821-06e4c11590b1" include("../ext/BATCubaExt.jl")
+        @require HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f" include("../ext/BATHDF5Ext.jl")
+        @require NestedSamplers = "41ceaf6f-1696-4a54-9b49-2e7a9ec3782e" include("../ext/BATNestedSamplersExt.jl")
+        @require Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80" include("../ext/BATPlotsExt.jl")
+        @require UltraNest = "6822f173-b0be-4018-9ee2-28bf56348d09" include("../ext/BATUltraNestExt.jl")
+    end
 end
 
 end # module
