+++ src/algodefaults/algodefaults.jl
@@ -8,3 +8,4 @@ include("default_autocor_len.jl")
 include("default_eff_sample_size.jl")
 include("default_mode_estimator.jl")
 include("default_integration_algorithm.jl")
+include("default_io_algorithm.jl")
