+++ src/algotypes/algotypes.jl
@@ -11,3 +11,4 @@ include("mode_estimator.jl")
 include("median_estimator.jl")
 include("differentiation_algorithm.jl")
 include("integration_algorithm.jl")
+include("io_algorithm.jl")
