+++ src/extdefs/extdefs.jl
@@ -1,0 +1,5 @@
+# This file is a part of BAT.jl, licensed under the MIT License (MIT).
+
+include("cuba_defs.jl")
+include("nestedsamplers_defs.jl")
+include("ultranest_defs.jl")
