--- src/io/io.jl
@@ -1,6 +1,0 @@
-# This file is a part of BAT.jl, licensed under the MIT License (MIT).
-
-include("hdf5.jl")
-include("bat_io.jl")
-
-# Included via require: "hdf5_specific.jl"
