+++ src/plotting/plotting.jl
@@ -4,6 +4,15 @@ const default_credibilities = [0.683, 0.955, 0.997]
 const default_colors = [:chartreuse2, :yellow, :red]
 
 
+# Implemented in BATPlotsExt:
+function _Plots_backend end
+function _Plots_cgrad end
+function _Plots_grid end
+function _Plots_Shape end
+function _Plots_Surface end
+function _Plots_backend_is_pyplot end
+
+
 include("MarginalDist.jl")
 include("recipes_MarginalDist_1D.jl")
 include("recipes_MarginalDist_2D.jl")
