--- src/samplers/nested_sampling/nested_sampling.jl
@@ -1,3 +1,0 @@
-# This file is a part of BAT.jl, licensed under the MIT License (MIT).
-
-# Included via @require: "ultranest.jl"
