+++ src/samplers/samplers.jl
@@ -4,4 +4,3 @@ include("bat_sample.jl")
 include("mcmc/mcmc.jl")
 include("sampled_density.jl")
 include("importance/importance_sampler.jl")
-include("nested_sampling/nested_sampling.jl")
