+++ test/Project.toml
@@ -23,7 +23,7 @@ NestedSamplers = "41ceaf6f-1696-4a54-9b49-2e7a9ec3782e"
 PDMats = "90014a1f-27ba-587c-ab20-58faa44d9150"
 ParallelProcessingTools = "8e8a01fc-6193-5ca1-a2f1-20776dae4199"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
-PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
+#PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
 QuadGK = "1fd47b50-473d-5c70-9696-f719f8f3bcdc"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Random123 = "74087812-796a-5b5d-8853-05524746bad3"
@@ -34,7 +34,7 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 TerminalLoggers = "5d786b92-1e48-4d6f-9151-6b4477ca9bed"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 TypedTables = "9d95f2ec-7b3d-5a63-8d20-e2491e220bb9"
-UltraNest = "6822f173-b0be-4018-9ee2-28bf56348d09"
+#UltraNest = "6822f173-b0be-4018-9ee2-28bf56348d09"
 UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 ValueShapes = "136a8f8c-c49b-4edb-8b98-f3d64d48be8f"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
